9063: abstracted the cloudwatch query-ing into its own helper
