+++ scripts/jest-scripts.config.ts
@@ -12,6 +12,7 @@ const config: Config = {
     '!circleci/*.ts',
     '!circleci/judge/bulkImportJudgeUsers.helpers.ts',
     '!circleci/judge/bulkImportJudgeUsers.ts',
+    '!cloudwatch/perform-query.ts',
     '!compareTypescriptErrors.ts',
     '!coverage/**',
     '!download-all-case-documents.ts',
@@ -25,6 +26,7 @@ const config: Config = {
     '!elasticsearch/health-migration.ts',
     '!elasticsearch/ready-cluster-for-migration.ts',
     '!elasticsearch/reindex.ts',
+    '!elasticsearch/retry-ocr-failures.ts',
     '!email/**',
     '!generate-uuid.ts',
     '!glue/**',
